feat(workflows): workflow embed viewer for the extension deep surface
- Extract RunDetail (+GateCard/RunTimeline/RunGraphSteps/StatusPill) out of
  WorkflowsPanel into RunDetail.tsx so both the sidebar panel and the embed
  iframe can mount it; WorkflowsPanel re-imports, behavior unchanged.
- EmbedWorkflowClient: two-tab (Runs / Edit) viewer reusing RunDetail +
  WorkflowBuilder, session-authed via the embed layout's fetch wrapper.
- embed/content/[id]: render workflow content type (was EmbedFallback);
  ?run= deep-links into a run's detail on the Runs tab.

Phase 0 gates green: typecheck, lint (0 err), build.
